Fix pathname canonicalization for input with non-standard protocols.
The `test()` and `exec()` methods also perform canonicalization on their
input on a component-by-component basis when the dictionary is passed.
In those cases we must be careful to use the correct type of pathname
canonicalization based on the protocol value.  Special schemes like
'https' get more typical encoding, but schemes like 'data' get a less
strict form of encoding.

This change also cleans up some fallback strings in `test()` and
`exec()` that are no longer needed.