Add "types" to package exports for TypeScript NodeNext
TypeScript 4.7 has support for package exports, however it requires that each package export entry specifies the typings for that export. If omitted, for a project that is using the new `NodeNext` resolution mode, the typings will not be found: https://devblogs.microsoft.com/typescript/announcing-typescript-4-7/#package-json-exports-imports-and-self-referencing

This PR adds a "types" field to the two export conditions, so that they can be imported when using the `NodeNext` setting in TypeScript 4.7.